backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.2.0",
    "np>=1.0.2",
    "ollama>=0.4.7",
    "openai>=1.63.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.5",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tabulate>=0.9.0",
]
